+++ friends/src/App.js
@@ -1,35 +1,8 @@
 import React, { Component } from "react";
 import "./App.css";
 import Axios from "axios";
-import { withStyles } from "@material-ui/core/styles";
-import Table from "@material-ui/core/Table";
-import TableBody from "@material-ui/core/TableBody";
-import TableCell from "@material-ui/core/TableCell";
-import TableHead from "@material-ui/core/TableHead";
-import TableRow from "@material-ui/core/TableRow";
-import Paper from "@material-ui/core/Paper";
-
-const styles = theme => ({
-  head: {
-    backgroundColor: theme.palette.common.black,
-    color: theme.palette.common.white
-  },
-  root: {
-    width: "70%",
-    marginTop: theme.spacing.unit * 3,
-    overflowX: "auto",
-    margin: "0 auto"
-  },
-  table: {
-    minWidth: 700
-  },
-
-  row: {
-    "&:nth-of-type(odd)": {
-      backgroundColor: "#EEEEEE",
-    }
-  }
-});
+import FriendsList from './components/FriendsList'
+import {Route } from 'react-router-dom'
 
 
 class App extends Component {
@@ -61,33 +34,9 @@ class App extends Component {
 
     return <div className="App">
         <h1>Friends</h1>
-        <Paper className={classes.root}>
-        <Table className={classes.table}>
-            <TableHead>
-              <TableRow>
-                <TableCell className={classes.head}>Name</TableCell>
-              <TableCell className={classes.head} numeric>Age</TableCell>
-              <TableCell className={classes.head}>Email</TableCell>
-              </TableRow>
-            </TableHead>
-
-            <TableBody>
-              {this.state.friends.map(friend => {
-                return <TableRow className={classes.row} key={friend.id}>
-                    <TableCell component="th" scope="row">
-                      {friend.name}
-                    </TableCell>
-                    <TableCell numeric>{friend.age}</TableCell>
-                    <TableCell component="th" scope="row">
-                      {friend.email}
-                    </TableCell>
-                  </TableRow>;
-              })}
-            </TableBody>
-          </Table>
-        </Paper>
+      <Route path="/" render={() => <FriendsList friends={this.state.friends} />} />
       </div>;
   }
 }
 
-export default withStyles(styles)(App);
+export default App;
